requirements.txt
gradio
requests
pillow
opencv-python
scikit-learn
numpy
scipy
pandas
svgwrite
vtracer